Add a Callback function when PXP is done
Some of our code runs the PXP from an interrupt.

For example when the camera completes reading in a frame
We then do the pixel conversion and start up a TFT screen update with the new data.

Not sure you want the cade to hang in the ISR waiting for the PXP operation to complete.  So instead of calling PXP_finish,

we can instead setup for a callback from the ISR, that we then start up the screen update.